Fix getkernel for fedora and fix missing packages
Fedora has kernels with a specific full name, so this needs to account
for those types. we still favour semver if possible so in case of
finding more than one kernel we can choose the higher one, but we allow
plain custom kernel names as well.

It also splits the services to apply to different OSes

Signed-off-by: Itxaka <itxaka@kairos.io>