Produce proper standard images with k3s/k0s versions
USes a provider flag and a defualt empty version flag to provide the
specific version that you want.

It goes directly to k0s installer and generates the service files
itself so it doesnt need to run anything

For k3s it uses the installers with some flags to generate the services
directly.

Signed-off-by: Itxaka <itxaka@kairos.io>