Drop variant flag
We dont need to have a variant flag, it wil be set automatically based
on the k8s provider.

If the value is empty, which is the default now, it will be a core image
as usual.

If any valid provider is set, then it will generate a standard image.

But we no longer need to set that flag manually thanks to cobra and its
proper defaults and such

Signed-off-by: Itxaka <itxaka@kairos.io>