Allow overrides per binary
This allows loading overrides for the binaries so during runtime you can
manipulate those.

Useful for testing new versions without having to wait for init to be
updated and released to bundle new versions. Bundled versions are still
there and they are the default.

Also reworks a bit the dockerfile to take advantage of docker cache on
rebuilds.

Reworks a bit the Makefile for the renaming of upstream releases and
repos so all its more standard now

Signed-off-by: Itxaka <itxaka@kairos.io>