Workaround having the proper modules to support http efi uki boot
This modules are needed to enqable transparent mounting of an iso from
which we booted via http booting. EFI will map and pass those memory
regions into userspace and the kernel will see that and try to mount it
as if it was a normal iso.

This allows us to get only the modules needed without having to installt
he full extra-modules pacakge which increases the size by 100Mb.

Signed-off-by: Itxaka <itxaka@kairos.io>