Enable networkd in rhel if its installed
It has come to our attention that some rhel users install the fedora
epel and use that to enable systemd-networkd.

As the default rhel repos do not have it, we dont enable it.

This patch makes it so if the systemd-networkd binary is available we
add it to the initramfs and to the enabled services so users can simple
add the epel repo and install systemd-networkd and they will get the
full benefits of it

It also explicitely install xz package as otherwise it will be removed
during the dracut packages cleanup as nothing installed it explicitely
so it wont be in the final image

Signed-off-by: Itxaka <itxaka@kairos.io>