Remove python3-pynvim from Debian family base packages
This package was unnecessary as it provides Python plugin support for Neovim
which is not required for base Kairos images. Additionally, python3-pynvim
pulls in the entire Python3 runtime as a dependency, increasing the image
size without providing essential functionality for the base system.

Users who need Python plugin support for Neovim can install it manually
when needed.

Signed-off-by: imusmanmalik <usmanmalik@engineer.com>