Improve mkinitfs under alpine
On the first install mkinitfs will create an automatic initramfs due to
its package trigger.

This initramfs is useless to us as it doesnt contain what we need on it,
so we always end up removing it afterwards.

Instead, just skip the trigger vi setting it into the mkinitfs config
file before installing and then restore the package mkinifs config file
to its proper place. This should save some time while installing and not
waste time into creating uneeded files

Signed-off-by: Itxaka <itxaka@kairos.io>