feat: Update Kubernetes provider configuration
Refactor the Kubernetes provider configuration in the init image.
This change includes the removal of the old provider enum and
introduces a new string field for the provider, along with a
configuration file option. Additionally, the installation stages
for Kubernetes have been removed, and a new event for provider
build installation has been added to streamline the process.

Signed-off-by: Itxaka <itxaka@kairos.io>